requirements.txt
fastapi
uvicorn
openai
python-dotenv
jinja2
python-multipart
numpy
google-api-python-client
google-auth
google-auth-oauthlib
dateparser
pytz
pytest
pytest-mock
pytest-asyncio
itsdangerous
sqlalchemy
alembic
slowapi